Added reading and loading file to common
